% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_interactions.R
\name{iv_gmm_fit}
\alias{iv_gmm_fit}
\title{GMM for Estimation of 2^K Factorial Design}
\usage{
iv_gmm_fit(y, d, z, method, max_iter = 15, tol = 1e-06)
}
\arguments{
\item{y}{vector of outcomes.}

\item{d}{K-column matrix of binary treatments.}

\item{z}{K-column matrix of binary instruments.}

\item{method}{character indiciating if the GMM should be
\code{"onestep"} using the identity weighting matrix (default) or
\code{"iterative"} to estimate the GMM weighting matrix.}

\item{max_iter}{maximum number of iterations to conduct when
estimating the iterative GMM.}

\item{tol}{criteria for stopping the iterative GMM procedure.}
}
\value{
A list of class \code{iv_factorial} that contains the following
  components: 
\item{rho}{vector of estimated compliance class
  probabilities.}
\item{psi}{vector of the estimated conditional mean of the outcome
  within the compliance classes.}
\item{init}{list of two vectors, with the initial starting values
  for the GMM estimation based on a just-identified estimation.}
\item{vcov}{estimated asymptotic variance matrix of the combined
  \code{rho} and \code{psi} parameters.}
\item{tau}{vector of estimated main effects of each factor among
  supercompliers.}
\item{tau_se}{vector of estimated standard errors for the
  estimated effects in \code{tau}.}
}
\description{
Fits a GMM for IV in 2^K Factorial Experiments
}
\author{
Matt Blackwell
}
