% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite-sample.R
\name{tidy.iv_finite_factorial}
\alias{tidy.iv_finite_factorial}
\title{Tidy an iv_finite_factorial object}
\usage{
\method{tidy}{iv_finite_factorial}(x, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An \code{iv_factorial} object produced by a call to
\code{\link[=iv_finite_factorial]{iv_finite_factorial()}}}

\item{conf.level}{The confidence level to use for the confidence
interval. Must be strictly greater than 0 and less than 1.
Defaults to  0.95, which corresponds to a 95
percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only.}
}
\value{

}
\description{
Tidy summarizes information about the components of a model.
}
\author{
Matt Blackwell
}
